frontend/requirements.txt
streamlit
requests
PyPDF2
python-docx
python-pptx
python-dotenv